Pull query iterator to the package level
